Added theme customizations to the PHP options
Custom themes can now be configured/used, by adding them to the captcha options. For each theme, you can set the icon color (dark or light, or more if you add them) and color of the border between the icons (as RGB values). The default themes can also be overwritten.

Note: You would still need to create the theme in CSS for it to work. This is only to configure it, to make it actually be accepted by the code.

- Added namespace to the classes.
- Fixed indentation of the ajax-form.php file.