Fixed bug which sometimes broke challenge generation on PHP 8.1
Sometimes on PHP 8.1, the challenge image would not generate and instead show a broken image. This bug was caused by a deprecation warning, which is thrown when floats are implicitly converted to integers (deprecated: Implicit conversion from float to int loses precision). It only occurred when generating challenges with uneven icon amounts.