Added latency compensation to timeout timestamps
The attempts cooldown and completion expiration timestamps now take the request latency into account. As the timestamps are generated on the server, they start 'counting down' even though the response might not have been received yet by the visitor. By adding the previous request latency onto the timestamp, it sort of compensates for the loss of time during the request.