Converted all icons to using true color
Note: A quick fix for new installations of IconCaptcha.

As pointed out in several issues and a pull request (https://github.com/fabianwennink/IconCaptcha-Plugin-jQuery-PHP/pull/28), there is a strange issue with the GD `imagerotate` function, which sometimes breaks on specific environments or versions. This appears to be a bug in the GD extension, but this is not confirmed.

Using `imagepalettetotruecolor` prior to rotating the icons seems to solve this issue, as pointed out in the pull request. As it's still being decided if imagepalettetotruecolor should be used, this is a quick fix for new installations using only the default icons.